Django/pyproject.toml
[tool.poetry]
name = "d-Party"
version = "0.1.0"
description = ""
authors = ["Takashi Uno <euno.eng@gmail.com>"]
readme = "README.md"
packages = [{include = "d_party"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.0"
channels = "^3.0.4"
gunicorn = "^20.1.0"
mysqlclient = "^2.1.0"
channels-redis = "^3.3.1"
uvicorn = {extras = ["standard"], version = "^0.16.0"}
django-debug-toolbar = "^3.2.4"
django-jazzmin = "^2.4.0"
pydantic = "^1.9.0"
django-boost = "^2.0"
djangorestframework = "^3.13.1"
django-filter = "^21.1"
djangochannelsrestframework = "^0.3.0"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-django = "^4.5.2"
black = "^21.12b0"
pytest-asyncio = "^0.18.1"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py310']